Added testsuite classes for PLR/PLA and LRR/LRA only, keeping the all methods included client/server ones -all tests passing-, explaining immediate and deferred LCS referencing 3GPP specs, plus some esthetic enhancements, also in SLh testsuite. Issues #37 and #38, PR #111
